requirements.txt
google_genai
simpleaudio